+++ CHANGELOG.md
@@ -1,2 +1,17 @@
+## 12.13.0 - 2026-03-07
+### Added
+- fwupd firmware updater
+- Flatpak settings
+- Graphical boot
+- KDE Connect
+- KDE Partition Manager
+### Changed
+- Move to Debian 12 and KDE Plasma 5
+- Featherpad as default editor
+- Firewalld instead of ufw
+### Removed
+- 32-bit/legacy support
+- VDPAU modules
+
 ## 11.11.0 - 2026-03-05
 ### Initial release